Minor change that allows arrays to be given as arguments in Disqus calls, in addition to strings.
Arrays are treated as multiple values for the same argument, to be used with Disqus method arguments that allow for multiple values. Such method arguments are marked in the API docs with 'allows multiple'.

I could not find an existing way to pass multiple values for the same argument using the PHP API, so I made this change. In any case, the change is fairly intuitive I think.

E.g.
  "thread" : array("1234", "2345")

is unrolled as "&thread=1234&thread=2345"